pyproject.toml
[project]
name = "retailco"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
